+++ internal/common/basic.go
@@ -8,6 +8,7 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
+	capiexp "sigs.k8s.io/cluster-api/exp/api/v1beta1"
 	cr "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/cluster-test-suites/internal/state"
@@ -74,28 +75,28 @@ func runBasic() {
 		})
 
 		It("has all its Deployments Ready (means all replicas are running)", func() {
-			Eventually(wait.Consistent(checkAllDeploymentsReady(wcClient), 10, time.Second)).
+			Eventually(wait.Consistent(CheckAllDeploymentsReady(wcClient), 10, time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
 		})
 
 		It("has all its StatefulSets Ready (means all replicas are running)", func() {
-			Eventually(wait.Consistent(checkAllStatefulSetsReady(wcClient), 10, time.Second)).
+			Eventually(wait.Consistent(CheckAllStatefulSetsReady(wcClient), 10, time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
 		})
 
 		It("has all its DaemonSets Ready (means all daemon pods are running)", func() {
-			Eventually(wait.Consistent(checkAllDaemonSetsReady(wcClient), 10, time.Second)).
+			Eventually(wait.Consistent(CheckAllDaemonSetsReady(wcClient), 10, time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
 		})
 
 		It("has all of its Pods in the Running state", func() {
-			Eventually(wait.Consistent(checkAllPodsSuccessfulPhase(wcClient), 10, time.Second)).
+			Eventually(wait.Consistent(CheckAllPodsSuccessfulPhase(wcClient), 10, time.Second)).
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(Succeed())
@@ -108,6 +109,22 @@ func runBasic() {
 				WithTimeout(15 * time.Minute).
 				WithPolling(wait.DefaultInterval).
 				Should(BeTrue())
+		})
+
+		It("has all machine pools ready and running", func() {
+			mcClient := state.GetFramework().MC()
+			cluster := state.GetCluster()
+
+			machinePools, err := state.GetFramework().GetMachinePools(context.Background(), cluster.Name, cluster.GetNamespace())
+			Expect(err).NotTo(HaveOccurred())
+			if len(machinePools) == 0 {
+				Skip("Machine pools are not found")
+			}
+
+			Eventually(wait.Consistent(CheckMachinePoolsReadyAndRunning(mcClient, cluster.Name, cluster.GetNamespace()), 5, 5*time.Second)).
+				WithTimeout(30 * time.Minute).
+				WithPolling(wait.DefaultInterval).
+				Should(Succeed())
 		})
 	})
 }
@@ -157,7 +174,7 @@ func CheckWorkerNodesReady(wcClient *client.Client, values *application.ClusterV
 	}
 }
 
-func checkAllPodsSuccessfulPhase(wcClient *client.Client) func() error {
+func CheckAllPodsSuccessfulPhase(wcClient *client.Client) func() error {
 	return func() error {
 		podList := &corev1.PodList{}
 		err := wcClient.List(context.Background(), podList)
@@ -178,7 +195,7 @@ func checkAllPodsSuccessfulPhase(wcClient *client.Client) func() error {
 	}
 }
 
-func checkAllDeploymentsReady(wcClient *client.Client) func() error {
+func CheckAllDeploymentsReady(wcClient *client.Client) func() error {
 	return func() error {
 		deploymentList := &appsv1.DeploymentList{}
 		err := wcClient.List(context.Background(), deploymentList)
@@ -200,7 +217,7 @@ func checkAllDeploymentsReady(wcClient *client.Client) func() error {
 	}
 }
 
-func checkAllStatefulSetsReady(wcClient *client.Client) func() error {
+func CheckAllStatefulSetsReady(wcClient *client.Client) func() error {
 	return func() error {
 		statefulSetList := &appsv1.StatefulSetList{}
 		err := wcClient.List(context.Background(), statefulSetList)
@@ -222,7 +239,7 @@ func checkAllStatefulSetsReady(wcClient *client.Client) func() error {
 	}
 }
 
-func checkAllDaemonSetsReady(wcClient *client.Client) func() error {
+func CheckAllDaemonSetsReady(wcClient *client.Client) func() error {
 	return func() error {
 		daemonSetList := &appsv1.DaemonSetList{}
 		err := wcClient.List(context.Background(), daemonSetList)
@@ -240,6 +257,55 @@ func checkAllDaemonSetsReady(wcClient *client.Client) func() error {
 		}
 
 		logger.Log("All (%d) daemonSets have all daemon pods running", len(daemonSetList.Items))
+		return nil
+	}
+}
+
+// CheckMachinePoolsReadyAndRunning checks if all MachinePool resources have Ready condition with Status True and are in
+// Running phase.
+func CheckMachinePoolsReadyAndRunning(mcClient *client.Client, clusterName string, clusterNamespace string) func() error {
+	return func() error {
+		machinePools := &capiexp.MachinePoolList{}
+		machinePoolListOptions := []cr.ListOption{
+			cr.InNamespace(clusterNamespace),
+			cr.MatchingLabels{
+				"cluster.x-k8s.io/cluster-name": clusterName,
+			},
+		}
+		err := mcClient.List(context.Background(), machinePools, machinePoolListOptions...)
+		if err != nil {
+			return err
+		}
+
+		if len(machinePools.Items) == 0 {
+			logger.Log("MachinePools not found.")
+			return nil
+		}
+
+		allMachinePoolsAreReadyAndRunning := true
+		for _, mp := range machinePools.Items {
+			machinePool := mp
+			var machinePoolIsReady bool
+			machinePoolIsReady, err = wait.IsClusterApiObjectConditionSet(&mp, capi.ReadyCondition, corev1.ConditionTrue, "")
+			if err != nil {
+				return err
+			}
+			allMachinePoolsAreReadyAndRunning = allMachinePoolsAreReadyAndRunning && machinePoolIsReady
+
+			currentMachinePoolPhase := capiexp.MachinePoolPhase(machinePool.Status.Phase)
+			machinePoolIsRunning := currentMachinePoolPhase == capiexp.MachinePoolPhaseRunning
+			allMachinePoolsAreReadyAndRunning = allMachinePoolsAreReadyAndRunning && machinePoolIsRunning
+			logger.Log(
+				"MachinePool '%s/%s' expected to be in Running phase, found MachinePool is in '%s' phase.",
+				machinePool.Namespace,
+				machinePool.Name,
+				machinePool.Status.Phase)
+		}
+
+		if !allMachinePoolsAreReadyAndRunning {
+			return fmt.Errorf("not all MachinePools are ready and running")
+		}
+
 		return nil
 	}
 }